Add temporary grade recalculation to MasterService(should be moved to Quartz). Fix average grade displaying on master page.

+++ src/main/java/io/mathlina/beautysalon/domain/Master.java
@@ -20,7 +20,7 @@ import org.springframework.data.domain.Pageable;
 @Entity
 public class Master extends User {
 
-  private Float grade;
+  private double grade;
 
   @ManyToMany(fetch = FetchType.LAZY)
   @JoinTable(
